% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection45dLine.R
\name{addReflectionPath}
\alias{addReflectionPath}
\title{addReflectionPath}
\usage{
addReflectionPath(
  p,
  fun,
  x_start_point,
  steps = 10,
  col = "red",
  add_arrow = F,
  n_arrows = 5,
  arrow_size = 0.3
)
}
\arguments{
\item{p}{ggplot object}

\item{fun}{a function}

\item{x_start_point}{x start point}

\item{steps}{reflect times}

\item{col}{line color}

\item{add_arrow}{add arrows?}

\item{n_arrows}{number of arrows}

\item{arrow_size}{size of arrows}
}
\value{
ggplot object
}
\description{
addReflectionPath
}
