% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxplot.R
\name{doAlign}
\alias{doAlign}
\title{doAlign}
\usage{
doAlign(plots, end_spacing = 0.1)
}
\arguments{
\item{plots}{a list of ggplot objects}

\item{end_spacing}{extend the space in the end, as a percentage of the size of xlim()}
}
\description{
doAlign
}
