% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{gghist}
\usage{
gghist(x, bins = 30, density = F, color = "grey95", ...)
}
\arguments{
\item{x}{a vector}

\item{bins}{bins}

\item{density}{add density if TRUE}

\item{color}{}
}
\value{

}
\description{
gghist
}
