% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVector.R
\name{ggVector}
\alias{ggVector}
\title{ggVector}
\usage{
ggVector(x, type = "l")
}
\arguments{
\item{x}{a vector, or a data.frame of vectors}

\item{type}{"l" for line, "p" for points, "b" for boths}
}
\value{
a ggplot object
}
\description{
plot a vector, or a data.frame of vectors
}
