% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection45dLine.R
\name{makeReflectionLine}
\alias{makeReflectionLine}
\title{makeReflectionLine}
\usage{
makeReflectionLine(fun, x_start_point, steps = 10)
}
\arguments{
\item{fun}{a function}

\item{x_start_point}{x start point}

\item{steps}{how many steps}
}
\value{
a data.frame of reflection line
}
\description{
get reflection line data
}
