% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_textbook.R
\name{theme_textbook}
\alias{theme_textbook}
\title{theme_textbook}
\usage{
theme_textbook(...)
}
\arguments{
\item{...}{argument pass to theme_bw()}
}
\value{
a gg theme
}
\description{
theme_bw() + no grid + no box + no background
}
